Абстрактный класс Смарт
