+++ src/view/edit-task.js
@@ -1,6 +1,6 @@
 import {COLORS} from '../const.js';
 import {isTaskExpired, isTaskRepeating, humanizeTaskDueDate} from '../utils/task.js';
-import AbstractView from "./abstract.js";
+import SmartView from "./smart.js";
 
 const BLANK_TASK = {
   color: COLORS[0],
@@ -133,7 +133,7 @@ const createEditTaskTemplate = (data) => {
   </article >`;
 };
 
-export default class TaskEdit extends AbstractView {
+export default class TaskEdit extends SmartView {
   constructor(task = BLANK_TASK) {
     super();
     this._data = TaskEdit.parseTaskToData(task);
@@ -189,33 +189,6 @@ export default class TaskEdit extends AbstractView {
     delete data.isRepeating;
 
     return data;
-  }
-
-  updateElement() {
-    let prevElement = this.getElement();
-    const parent = prevElement.parentElement;
-    this.removeElement();
-
-    const newElement = this.getElement();
-
-    parent.replaceChild(newElement, prevElement);
-    prevElement = null; // Чтобы окончательно "убить" ссылку на prevElement
-
-    this.restoreHandlers();
-  }
-
-  updateData(update, justDataUpdating) {
-    if (!update) {
-      return;
-    }
-
-    this._data = Object.assign({}, this._data, update);
-
-    if (justDataUpdating) {
-      return;
-    }
-
-    this.updateElement();
   }
 
   _dueDateToggleHandler(evt) {
